[chore] refactor AMP validation
