+++ .changeset/poor-lizards-punch.md
@@ -1,0 +1,5 @@
+---
+'@sveltejs/kit': patch
+---
+
+[chore] refactor AMP validation
